sample /tasks/main.yml
sample file in directory [ansible roles/ app/tasks/main.yml] for multi-tier architecture:
Here is an example of the main.yml file in the tasks directory for the app role in a multi-tier architecture:

This file specifies the tasks that will be executed by the app role. The tasks are executed in the order they are listed. The tasks include installing required packages, creating the application directory, deploying the application code, configuring the firewall, restarting the firewall service, and starting the application service.

Each task is defined using the - name: syntax, followed by the task name, and the task action is defined using the module syntax, such as apt, file, copy, template, and service. The become: yes directive is used to execute the task as a privileged user, typically root.

This main.yml file provides a blueprint for deploying the application component in a multi-tier architecture using Ansible. The variables used in the file, such as packages, app_path, app_user, app_group, app_code, app_service, etc., can be defined in a separate variables file and included in the playbook using the vars_files directive.