sample /templates/app.conf.j2
sample file in directory [ansible roles/ app/templates/app.conf.j2] for multi-tier architecture:
The file app.conf.j2 in the directory [ansible roles/ app/templates/app.conf.j2] might contain a configuration template for the application in a multi-tier architecture. The exact contents of the template would depend on the specifics of the architecture and the application being deployed. However, here is an example of what a sample template might look like:

In this example, the template includes variables for the port the application should listen on, the hostname or IP address of the database server, the port for the database server, the database username, and the database password. These variables would be populated by Ansible when the template is used. The exact implementation will depend on the specifics of the architecture and the application being deployed.